% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComparacaoMedias.R
\name{ComparacaoMedias}
\alias{ComparacaoMedias}
\title{Comparacao de medias}
\usage{
ComparacaoMedias(y, trt, DFerror, MSerror, alpha = 0.05, group =
 TRUE,main = NULL)
}
\arguments{
\item{y}{Vetor numerico contendo a variavel resposta.}

\item{trt}{Vetor numerico ou complexo contendo os tratamentos.}

\item{DFerror}{Grau de liberdade do residuo.}

\item{MSerror}{Quadrado medio do residuo.}

\item{alpha}{Nivel nominal de significancia.}

\item{group}{TRUE ou FALSE}

\item{main}{Titulo.}
}
\value{
Retorna a comparacao multipla de medias obtida por varios testes.
}
\description{
"This function returns the multiple
 comparison of means by the t, t with Bonferroni protection, Duncan, SNK,
 Tukey and Scott-Knott tests".
 Esta funcao retorna a comparacao multipla de medias obtidas pelos
 testes t, t com protecao de Bonferroni, Duncan, SNK, Tukey e Scott-Knott
 (Funcao adaptada do expDes.pt).
}
\references{
BANZATTO, D. A.; KRONKA, S. N. Experimentacao Agricola. 4 ed.
 Jaboticabal: Funep. 2006. 237 p. ISBN: 85-87632-71-X

 GOMES, F. P. Curso de Estatistica Experimental. 10a ed. Piracicaba:
 ESALQ/USP. 1982. 430.
}
