% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados1.R
\docType{data}
\name{Dados1}
\alias{Dados1}
\title{Dados de exemplo (1 fator e 1 testemunha).}
\format{
Um data frame com 20 observacoes das seguintes variaveis. \describe{
  \item{\code{Test}}{Vetor do tipo Fator.} \item{\code{Espacamento}}{Vetor do
  tipo Fator.} \item{\code{Rep}}{Vetor numerico.} \item{\code{Prod}}{Vetor
  numerico.} }
}
\usage{
data(Dados1)
}
\description{
Exemplo simulado de um experimento com um fator qualitativo, uma
  testemunha adicional e 4 repeticoes. "Simulated example of an experiment
  with an qualitative factor, additional control and 4 repetitions".
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{<alcineimistico@hotmail.com>}
}
\keyword{datasets}
