% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados2.R
\docType{data}
\name{Dados2}
\alias{Dados2}
\title{Dados de exemplo (1 fator e 3 testemunhas).}
\format{
Um data frame com 28 observacoes das seguintes variaveis. \describe{
  \item{\code{Test}}{Vetor do tipo Fator.} \item{\code{Dose}}{Vetor do tipo
  numerico.} \item{\code{Rep}}{Vetor numerico.} \item{\code{Prod}}{Vetor
  numerico.} }
}
\usage{
data(Dados2)
}
\description{
Exemplo simulado de um experimento com um fator quantitativo e
  tres testemunhas adicionais e 4 repeticoes. "Simulated example of an
  experiment with an quantitative factor, 3 additional controls and 4
  repetitions".
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{<alcineimistico@hotmail.com>}
}
\keyword{datasets}
