% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados3.R
\docType{data}
\name{Dados3}
\alias{Dados3}
\title{Dados de exemplo (2 fatores e 2 testemunhas).}
\format{
Um data frame com 28 observacoes das seguintes variaveis. \describe{
  \item{\code{Testemunha}}{Vetor do tipo Fator.} \item{\code{FatorA}}{Vetor
  do tipo Fator.} \item{\code{FatorB}}{Vetor do tipo Fator.}
  \item{\code{Rep}}{Vetor do tipo numerico} \item{\code{NumeroFrutos}}{Vetor
  numerico.} }
}
\usage{
data(Dados3)
}
\description{
Exemplo simulado de um experimento com dois fatores
  qualitativos, duas testemunhas adicionais e duas repeticoes."Simulated
  example of an experiment with an 2 qualitative factors, additionals
  controls and 4 repetitions".
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
