% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados4.R
\docType{data}
\name{Dados4}
\alias{Dados4}
\title{Dados de exemplo (2 fatores e 1 testemunha).}
\format{
Um data frame com 36 observacoes das seguintes variaveis. \describe{
  \item{\code{Test}}{Vetor do tipo Fator.} \item{\code{Doses}}{Vetor do tipo
  quantitativo.} \item{\code{Cultivar}}{Vetor do tipo Fator.}
  \item{\code{Rep}}{Vetor do tipo numerico} \item{\code{Prod}}{Vetor
  numerico.} }
}
\usage{
data(Dados4)
}
\description{
Exemplo simulado de um experimento com dois fatores um
  quantitativo e o outro qualitativo, uma testemunha e quatro
  repeticoes."Simulated example of an experiment with an qualitative and
  quantitative factors, additional control and 4 repetitions".
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
