% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados5.R
\docType{data}
\name{Dados5}
\alias{Dados5}
\title{Dados de exemplo (2 fatores e 3 testemunha).}
\format{
Um data frame com 44 observacoes das seguintes variaveis. \describe{
\item{\code{Test}}{Vetor do tipo Fator.} \item{\code{Cultivar}}{Vetor do tipo
quantitativo.} \item{\code{Doses}}{Vetor do tipo Fator.}
\item{\code{Rep}}{Vetor do tipo numerico} \item{\code{Prod}}{Vetor numerico.}
}
}
\usage{
data(Dados5)
}
\description{
Exemplo simulado de um experimento com dois fatores um
  quantitativo e o outro qualitativo, uma testemunha e quatro
  repeticoes."Simulated example of an experiment with qualitative and
  quantitative factors, additional control and 4 repetitions".
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{<alcineimistico@hotmail.com>}
}
\keyword{datasets}
