% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados6.R
\docType{data}
\name{Dados6}
\alias{Dados6}
\title{Dados de exemplo (3 fatores e 3 testemunha).}
\format{
Um data frame com 76 observacoes das seguintes variaveis. \describe{
\item{\code{Test}}{Vetor do tipo Fator.} \item{\code{Cultivar}}{Vetor do tipo
quantitativo.} \item{\code{Espacamento}}{Vetor do tipo Fator.}
\item{\code{Adubacao}}{Vetor do tipo Fator.} \item{\code{Rep}}{Vetor do tipo
numerico} \item{\code{Prod}}{Vetor numerico.} }
}
\usage{
data(Dados6)
}
\description{
Exemplo simulado de um experimento com tres fatores
  qualitativos, uma testemunha e quatro repeticoes. "Simulated example of an
  experiment with 3 qualitative factors, additional control
  and 4 repetitions".
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{<alcineimistico@hotmail.com>}
}
\keyword{datasets}
