% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados7.R
\docType{data}
\name{Dados7}
\alias{Dados7}
\title{Dados de exemplo de DQL (4 tratamentos comuns e 2 testemunhas).}
\format{
Um data frame com 36 observacoes das seguintes variaveis. \describe{
\item{\code{Test}}{Vetor do tipo Fator.}
 \item{\code{Medicamento}}{Vetor com os tratamentos.} \item{\code{Animal}}{Vetor com a identificao das linhas.}
\item{\code{Epoca}}{Vetor com a identificacao das colunas.}
\item{\code{Carrapato}}{Vetor do tipo numerico com a variavel resposta.}
}
}
\usage{
data(Dados7)
}
\description{
Exemplo simulado de um experimento em DQL com quatro tratamentos
  comuns qualitativos e duas testemunhas.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
