% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dados8.R
\docType{data}
\name{Dados8}
\alias{Dados8}
\title{Dados de exemplo de DQL (5 tratamentos comuns e 1 testemunha).}
\format{
Um data frame com 36 observacoes das seguintes variaveis. \describe{
\item{\code{Test}}{Vetor do tipo Fator.}
\item{\code{Medicamento}}{Vetor numerico com os tratamentos quantitativos.}
 \item{\code{Animal}}{Vetor com a identificao das linhas.}
\item{\code{Epoca}}{Vetor com a identificacao das colunas.}
\item{\code{Carrapato}}{Vetor do tipo numerico com a variavel resposta.}
}
}
\usage{
data(Dados8)
}
\description{
Exemplo simulado de um experimento em DQL com cinco tratamentos
  comuns quantitativos e uma testemunha.
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\author{
Alcinei Mistico Azevedo, \email{alcineimistico@hotmail.com}
}
\keyword{datasets}
