% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DunnetCritic.R
\name{DunnetCritic}
\alias{DunnetCritic}
\title{Valor Critico do teste Dunnet}
\usage{
DunnetCritic(gln,gld,sig)
}
\arguments{
\item{gln}{Numero de comparacoes a serem realizadas.}

\item{gld}{Numero de graus de liberdade do residuo.}

\item{sig}{Nivel de significancia, deve receber o valor 0.001, 0.01,
0.05 ou 0.01.}
}
\value{
Retorna o valor do teste Dunnett (bilateral).
}
\description{
Esta funcao retorna o valor critico do teste bilateral
 Dunnet."This function returns the critical value of the Dunnet bilateral
 test."
}
\references{
BANZATTO, D. A.; KRONKA, S. N. Experimentacao Agricola. 4 ed.
 Jaboticabal: Funep. 2006. 237 p.

 GOMES, F. P. Curso de Estatistica Experimental. 10a ed. Piracicaba:
 ESALQ/USP. 1982. 430.
}
