% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interpolacao.R
\name{Interpolacao}
\alias{Interpolacao}
\title{Interpolacao}
\usage{
Interpolacao(Mat,GLtrat,GLres,gln,gld)
}
\arguments{
\item{Mat}{Matriz contendo os valores tabelados.}

\item{GLtrat}{Vetor com os graus de liberdade correspondente as
colunas.}

\item{GLres}{Vetor com os graus de liberdade correspondente as linhas.}

\item{gln}{Graus de liberdade desejado na coluna.}

\item{gld}{Graus de liberdade desejado na linha.}
}
\value{
Retorna o valor tabelado apos a interpolacao.
}
\description{
Esta funcao faz a interpolacao para a obtencao de valores
 tabelados.
}
