% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressaoPolinomial.R
\name{RegressaoPolinomial}
\alias{RegressaoPolinomial}
\title{Regressao polinomial}
\usage{
RegressaoPolinomial(resp, trat, glres, SQres, gltrat, SQtrat,verbose=T)
}
\arguments{
\item{resp}{Vetor numerico contendo a variavel resposta.}

\item{trat}{Vetor contendo a identificacao dos tratamentos.}

\item{glres}{Grau de liberdade dos residuos.}

\item{SQres}{Soma de quadrado dos residuos.}

\item{gltrat}{Grau de liberdade de tratamentos.}

\item{SQtrat}{Soma de quadrado de tratamentos.}

\item{verbose}{Valor logico (TRUE/FALSE). TRUE apresenta os resultados da analise.}
}
\value{
Retorna os resultads da analise de regressao.
}
\description{
Esta funcao retorna os resultados da analise de regressao linear,
 quadratica e cubica (Funcao adaptada do expDes.pt). "This function returns
 the results of the linear, quadratic and cubic regression analysis."
}
\references{
BANZATTO, D. A.; KRONKA, S. N. Experimentacao Agricola. 4 ed.
 Jaboticabal: Funep. 2006. 237 p.

 GOMES, F. P. Curso de Estatistica Experimental. 10a ed. Piracicaba:
 ESALQ/USP. 1982. 430.
}
