% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tratamentos.ad-package.R
\docType{package}
\name{Tratamentos.ad-package}
\alias{Tratamentos.ad-package}
\alias{Tratamentos.ad}
\title{Tratamentos.ad: Pacote Para Analise De Experimentos Com Testemunhas}
\description{
Tratamentos.ad: realiza a analise de experimentos com unifatorial ou
  fatorial contendo uma ou mais testemunhas adicionais. Para os fatores
  quantitativos e feita a analise de regressao, ja para os fatores
  qualitativos sao feitos testes de comparacao multipla. Para comparar os
  tratamentos testemunhas com os demais e utilizado o teste Dunnett.
}
\details{
Este pacote realiza a analise de experimentos com unifatorial ou  fatorial
contendo uma ou mais testemunhas
}
\references{
Canal no youtube -> Alcinei Azevedo - Dicas e aulas
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>

 BANZATTO, D. A.; KRONKA, S. N. Experimentacao Agricola. 4 ed.
 Jaboticabal: Funep. 2006. 237 p.

 GOMES, F. P. Curso de Estatistica Experimental. 10a ed. Piracicaba:
 ESALQ/USP. 1982. 430.
}
\seealso{
\code{\link{dic.ad}}, \code{\link{dbc.ad}},
 \code{\link{fatorial2.ad.dic}}, \code{\link{fatorial2.ad.dbc}}
}
