% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tratamentos.ad-package.R
\docType{package}
\name{Tratamentos.ad-package}
\alias{Tratamentos.ad-package}
\alias{Tratamentos.ad}
\title{Tratamentos.ad: Pacote Para Analise De Experimentos Com Testemunhas}
\description{
Tratamentos.ad: realiza a analise de experimentos com unifatorial ou
  fatorial contendo uma ou mais testemunhas adicionais. Para os fatores
  quantitativos e feita a analise de regressao, ja para os fatores
  qualitativos sao feitos testes de comparacao multipla. Para comparar os
  tratamentos testemunhas com os demais e utilizado o teste Dunnett.
}
\details{
Este pacote realiza a analise de experimentos com unifatorial ou  fatorial
contendo uma ou mais testemunhas
}
\references{
<https://www.youtube.com/playlist?list=PLvth1ZcREyK4wSzwg-IxvrzaNzSLLrXEB>
}
\seealso{
{dic.ad,dbc.ad,fatorial2.dic.ad,fatorial2.dbc.ad}
}
